requirements.txt
numpy >= 1.22.2
discord.py >= 1.7.3
discord >= 2.1.0
matplotlib >= 3.6.3
japanize_matplotlib
lark
python-dotenv
pandas
fastapi
uvicorn
sympy